Refactor into KNNVectorFieldMapperUtil and split test into two tests
